consolidated menu.py into main.py, added error handling to text parser
