+++ Game/game.py
@@ -6,14 +6,16 @@ class Game:
 
 	def startGame(self):
 		while 1:
-			self.getInput()
+			args = input("Enter an action: ").lower().split()
+			self.getInput(args)
 
-	def getInput(self):
-		action = self.parser.parse()
+	def getInput(self, args):
+		action = self.parser.parse(args)
 	
 		if action == "quit":
 			print("Exiting gameplay")
 			exit()
 		else:
 			print(action)
+			print()
